chore: rename server -> app and update imports
